refactor: Move docpad and bottom menu to features
