+++ src/app/(frontend)/articles/[slug]/page.tsx
@@ -10,7 +10,7 @@ import { LivePreviewListener } from '@/components/LivePreviewListener'
 import { RelatedArticles } from '@/features/articles/components/RelatedArticles'
 import { ArticleTopMenuContainer } from '@/features/articles/components/ArticleTopMenuContainer'
 import { RightMenuContainer } from '@/features/shared/components/RightMenu'
-import BottomMenu from '@/components/BottomMenu'
+import { BottomMenu } from '@/features/shared/components/BottomMenu'
 import { ArticleHero } from '@/heros/ArticleHero'
 import { RenderArticleBlocks } from '@/blocks/RenderArticleBlocks'
 import { ArticleLeftMenuContainer } from '@/features/articles/components/ArticleLeftMenuContainer'
